Split smarthome/index.ts into discovery.ts, handlers.ts, and index.ts
discovery.ts holds ALARM_ENDPOINT_ID and buildDiscoveryEndpoints (the
Alexa device schema, shared by both inbound discovery requests and the
outbound sync). handlers.ts holds the individual request handlers and
response property builders. index.ts is now the thin public interface:
smarthomeHandlers, syncDiscovery, and sendSimpleEventSource.
